refactor: dateToYYMMDD 경로 변경 및 separator 추가
- shared/functions/date.ts에 정의
- separator를 이용해 원하는 포맷으로 정의 가능
